refactor test_batch_invariant to rlhf/consistency
